Fixed integration commands, added UART functionality, temporarily removed UDP
Part of a more general refactor of the code. You should be able to change the integration parameters for the device correctly now, and now you can communicate to the ESP32 over UART. I've added a few parameters to the Kconfig revolving around the UART, I may add more for the SPI as well.